+++ src/spring/HISTORY.md
@@ -10,6 +10,9 @@ Release History
 * Support route to app level in command `az spring gateway route-config create` and `az spring gateway route-config create`
 * Add new command group `az spring dev-tool` to manage Dev Tools.
 * Add argument `--enable-application-live-view` in command `az spring create` to support enable application live view when creating Enterprise sku Spring resource.
+* Add new command group `az spring application-accelerator customized-accelerator` to manage Customized Accelerator.
+* Add new command group `az spring application-accelerator predefined-accelerator` to manage Predefined Accelerator.
+* Add argument `--enable-application-accelerator` in command `az spring create` to support enable application accelerator when creating Enterprise sku Spring resource.
 
 1.2.0
 ---